+++ src/main/java/com/fot/backcontask/service/contract/ContractService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.contract;
 import com.fot.backcontask.model.Contract;
 import com.fot.backcontask.service.Service;
 
-public interface ContractService extends Service<Contract, Integer>{
+public interface ContractService extends Service<Contract, Long>{
 
 }
