+++ src/main/java/com/fot/backcontask/service/contracttype/ContractTypeService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.contracttype;
 import com.fot.backcontask.model.ContractType;
 import com.fot.backcontask.service.Service;
 
-public interface ContractTypeService extends Service<ContractType, Integer> {
+public interface ContractTypeService extends Service<ContractType, Long> {
 
 }
