+++ src/main/java/com/fot/backcontask/service/employee/EmployeeService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.employee;
 import com.fot.backcontask.model.Employee;
 import com.fot.backcontask.service.Service;
 
-public interface EmployeeService extends Service<Employee, Integer> {
+public interface EmployeeService extends Service<Employee, Long> {
 
 }
