+++ src/main/java/com/fot/backcontask/service/key/KeyService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.key;
 import com.fot.backcontask.model.Key;
 import com.fot.backcontask.service.Service;
 
-public interface KeyService extends Service<Key, Integer> {
+public interface KeyService extends Service<Key, Long> {
 
 }
