+++ src/main/java/com/fot/backcontask/service/keytype/KeyTypeService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.keytype;
 import com.fot.backcontask.model.KeyType;
 import com.fot.backcontask.service.Service;
 
-public interface KeyTypeService extends Service<KeyType, Integer> {
+public interface KeyTypeService extends Service<KeyType, Long> {
 
 }
