+++ src/main/java/com/fot/backcontask/service/population/PopulationService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.population;
 import com.fot.backcontask.model.Population;
 import com.fot.backcontask.service.Service;
 
-public interface PopulationService extends Service<Population, Integer> {
+public interface PopulationService extends Service<Population, Long> {
 
 }
