+++ src/main/java/com/fot/backcontask/service/province/ProvinceService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.province;
 import com.fot.backcontask.model.Province;
 import com.fot.backcontask.service.Service;
 
-public interface ProvinceService extends Service<Province, Integer>{
+public interface ProvinceService extends Service<Province, Long>{
 
 }
